Renamed directions-refresh/v1 to v5 and updated to run on production
