+++ samples/src/main/java/com/mapbox/samples/BasicDirectionsRefresh.java
@@ -2,8 +2,8 @@ package com.mapbox.samples;
 
 import com.mapbox.api.directions.v5.MapboxDirections;
 import com.mapbox.api.directions.v5.models.DirectionsResponse;
-import com.mapbox.api.directionsrefresh.v1.MapboxDirectionsRefresh;
-import com.mapbox.api.directionsrefresh.v1.models.DirectionsRefreshResponse;
+import com.mapbox.api.directionsrefresh.v5.MapboxDirectionsRefresh;
+import com.mapbox.api.directionsrefresh.v5.models.DirectionsRefreshResponse;
 import com.mapbox.geojson.Point;
 import com.mapbox.sample.BuildConfig;
 
@@ -15,6 +15,7 @@ import retrofit2.Response;
 
 import static com.mapbox.api.directions.v5.DirectionsCriteria.ANNOTATION_CONGESTION;
 import static com.mapbox.api.directions.v5.DirectionsCriteria.OVERVIEW_FULL;
+import static com.mapbox.api.directions.v5.DirectionsCriteria.PROFILE_DRIVING_TRAFFIC;
 
 public class BasicDirectionsRefresh {
   public static void main(String[] args) throws IOException {
@@ -29,11 +30,12 @@ public class BasicDirectionsRefresh {
       .origin(Point.fromLngLat(-95.6332, 29.7890))
       .destination(Point.fromLngLat(-95.3591, 29.7576))
       .overview(OVERVIEW_FULL)
+      .profile(PROFILE_DRIVING_TRAFFIC)
       .annotations(ANNOTATION_CONGESTION).build();
 
     Response<DirectionsResponse> response = directions.executeCall();
+    System.out.println("Directions response: " + response);
     String requestId = response.body().routes().get(0).routeOptions().requestUuid();
-    System.out.println("Directions response: " + response + " " + requestId);
 
     return requestId;
   }
