+++ services-directions-refresh/src/main/java/com/mapbox/api/directionsrefresh/v5/package-info.java
@@ -3,4 +3,4 @@
  *
  * @since 4.4.0
  */
-package com.mapbox.api.directionsrefresh.v1;
+package com.mapbox.api.directionsrefresh.v5;
